refactor: removed non necessary files after refactoring to simpler architecture
