--- src/repositories/courseRepository.ts
@@ -1,50 +1,0 @@
-import { isNullableType } from "graphql";
-import { FilterQuery } from "mongoose";
-import AppError from "../errors/AppError";
-import { Course } from "../models/Course";
-import { PaginatedResult } from "../utils/types";
-import Repository, { IRepository } from "./repository";
-
-export interface ICourseRepository extends IRepository<Course> {
-  create(code: string, name: string, description: string): Promise<Course>;
-
-  findAndUpdateCourse(
-    id: string,
-    name: string,
-    description: string
-  ): Promise<Course>;
-}
-
-class CourseRepository extends Repository<Course> implements ICourseRepository {
-  async create(
-    code: string,
-    name: string,
-    description: string
-  ): Promise<Course> {
-    const course = await this.model.create({
-      code,
-      name,
-      description,
-    });
-
-    return course;
-  }
-
-  async findAndUpdateCourse(
-    id: string,
-    name: string,
-    description: string
-  ): Promise<Course> {
-    const course = await this.model.findById(id);
-    if (!course) throw new AppError("Course not found", 404);
-
-    if (name && name.length > 3) course.name = name;
-    if (description && description.length > 5) course.description = description;
-
-    await course.save();
-
-    return course;
-  }
-}
-
-export default CourseRepository;
