+++ src/routes/courseRoutes.ts
@@ -1,7 +1,5 @@
-import CourseController from "../controllers/courseController";
 import { Router } from "express";
 import { deleteImageFromCloudinary, upload } from "../utils/cloudinaryUpload";
-import ServiceLocator from "../di/serviceLocator";
 import createCourse from "../controllers/course/createCourse";
 import getAllCourses from "../controllers/course/getAllCourses";
