+++ src/utils/cloudinaryUpload.ts
@@ -8,7 +8,6 @@ import dotenv from "dotenv";
 dotenv.config();
 
 const { CLOUD_NAME, API_KEY, API_SECRET } = process.env;
-console.log(CLOUD_NAME, API_KEY, API_SECRET);
 cloudinary.config({
   cloud_name: CLOUD_NAME,
   api_key: API_KEY,
